Move DiffKit logic to the correct files
