+++ Sources/StreamChatUI/ChatMessageList/ChatMessageListVC.swift
@@ -383,17 +383,6 @@ open class ChatMessageListVC: _ViewController,
         navigationController?.present(alert, animated: true)
     }
 
-    /// Set the previous message snapshot before the data controller reports new messages.
-    internal func setPreviousMessagesSnapshot(_ messages: [ChatMessage]) {
-        listView.previousMessagesSnapshot = messages
-    }
-
-    /// Set the new message snapshot reported by the data controller.
-    internal func setNewMessagesSnapshot(_ messages: [ChatMessage]) {
-        listView.currentMessagesFromDataSource = messages
-        listView.newMessagesSnapshot = messages
-    }
-
     // MARK: - UITableViewDataSource & UITableViewDelegate
 
     open func numberOfSections(in tableView: UITableView) -> Int {
